Scoll to Button added & JS code Refactoring
