+++ style.css
@@ -647,6 +647,14 @@ sup {
   position: sticky;
 }
 
+.right-0 {
+  right: 0px;
+}
+
+.bottom-0 {
+  bottom: 0px;
+}
+
 .top-0 {
   top: 0px;
 }
@@ -661,10 +669,6 @@ sup {
 
 .bottom-1 {
   bottom: 0.25rem;
-}
-
-.right-0 {
-  right: 0px;
 }
 
 .left-\[10px\] {
@@ -685,10 +689,6 @@ sup {
 
 .-left-10 {
   left: -2.5rem;
-}
-
-.bottom-0 {
-  bottom: 0px;
 }
 
 .top-40 {
@@ -717,6 +717,10 @@ sup {
 
 .left-\[90\%\] {
   left: 90%;
+}
+
+.z-30 {
+  z-index: 30;
 }
 
 .z-50 {
@@ -899,6 +903,14 @@ sup {
   margin-right: 10%;
 }
 
+.mr-6 {
+  margin-right: 1.5rem;
+}
+
+.mb-6 {
+  margin-bottom: 1.5rem;
+}
+
 .mt-7 {
   margin-top: 1.75rem;
 }
@@ -925,10 +937,6 @@ sup {
 
 .mt-\[10vh\] {
   margin-top: 10vh;
-}
-
-.mb-6 {
-  margin-bottom: 1.5rem;
 }
 
 .mb-10 {
@@ -1139,10 +1147,6 @@ sup {
   margin-right: 3.5rem;
 }
 
-.mr-6 {
-  margin-right: 1.5rem;
-}
-
 .mt-\[8px\] {
   margin-top: 8px;
 }
@@ -1305,6 +1309,10 @@ sup {
 
 .aspect-square {
   aspect-ratio: 1 / 1;
+}
+
+.h-8 {
+  height: 2rem;
 }
 
 .h-full {
@@ -1503,6 +1511,10 @@ sup {
   height: 0.25rem;
 }
 
+.h-6 {
+  height: 1.5rem;
+}
+
 .max-h-14 {
   max-height: 3.5rem;
 }
@@ -1521,6 +1533,10 @@ sup {
 
 .min-h-full {
   min-height: 100%;
+}
+
+.w-8 {
+  width: 2rem;
 }
 
 .w-full {
@@ -1719,6 +1735,10 @@ sup {
 
 .w-\[11\%\] {
   width: 11%;
+}
+
+.w-7 {
+  width: 1.75rem;
 }
 
 .min-w-full {
@@ -2195,6 +2215,10 @@ sup {
   white-space: nowrap;
 }
 
+.rounded-full {
+  border-radius: 9999px;
+}
+
 .rounded {
   border-radius: 0.25rem;
 }
@@ -2205,10 +2229,6 @@ sup {
 
 .rounded-lg {
   border-radius: 0.5rem;
-}
-
-.rounded-full {
-  border-radius: 9999px;
 }
 
 .rounded-md {
@@ -2411,6 +2431,11 @@ sup {
   border-left-color: rgb(249 250 251 / var(--tw-border-opacity));
 }
 
+.bg-blue-500 {
+  --tw-bg-opacity: 1;
+  background-color: rgb(59 130 246 / var(--tw-bg-opacity));
+}
+
 .bg-transparent {
   background-color: transparent;
 }
@@ -2558,11 +2583,6 @@ sup {
   background-color: rgb(255 255 255 / var(--tw-bg-opacity));
 }
 
-.bg-blue-500 {
-  --tw-bg-opacity: 1;
-  background-color: rgb(59 130 246 / var(--tw-bg-opacity));
-}
-
 .bg-pink-200 {
   --tw-bg-opacity: 1;
   background-color: rgb(251 207 232 / var(--tw-bg-opacity));
@@ -2646,6 +2666,10 @@ sup {
      object-position: center;
 }
 
+.p-2 {
+  padding: 0.5rem;
+}
+
 .p-0 {
   padding: 0px;
 }
@@ -2660,10 +2684,6 @@ sup {
 
 .p-5 {
   padding: 1.25rem;
-}
-
-.p-2 {
-  padding: 0.5rem;
 }
 
 .p-2\.5 {
@@ -3138,12 +3158,12 @@ sup {
   font-size: 32px;
 }
 
-.font-semibold {
-  font-weight: 600;
-}
-
 .font-bold {
   font-weight: 700;
+}
+
+.font-semibold {
+  font-weight: 600;
 }
 
 .font-medium {
@@ -3775,6 +3795,11 @@ sup {
 .hover\:border-orange-700:hover {
   --tw-border-opacity: 1;
   border-color: rgb(194 65 12 / var(--tw-border-opacity));
+}
+
+.hover\:bg-blue-700:hover {
+  --tw-bg-opacity: 1;
+  background-color: rgb(29 78 216 / var(--tw-bg-opacity));
 }
 
 .hover\:bg-blue-800:hover {
